third_party/blink/web_tests/external/wpt/editing/other/formatblock-preserving-selection.tentative_styleWithCSS=false_block=dd-expected.txt

This is a testharness.js-based test.
[FAIL] Preserve selection after formatBlock with dd at <table><tr><td>a[bc</td></tr></table><div>de]f</div>
  assert_equals: Result: <dd><table><tbody><tr><td><dd>abc</dd></td></tr></tbody></table>def</dd> expected "bcde" but got "bcdef"
Harness: the test ran to completion.
